kube-autofix/requirements.txt
kubernetes>=29.0.0,<31.0.0
openai>=1.30.0,<2.0.0
pydantic>=2.7.0,<3.0.0
pydantic-settings>=2.3.0,<3.0.0
pyyaml>=6.0.1,<7.0.0
rich>=13.7.0,<14.0.0
python-dotenv>=1.0.0,<2.0.0
tenacity>=8.3.0,<9.0.0
click>=8.1.0,<9.0.0
